An issue with minimum deployment
