+++ Chatty/Controller/ChatViewController.swift
@@ -8,96 +8,37 @@ import UIKit
 import FirebaseAuth
 import FirebaseFirestore
 
-class ChatViewController: UIViewController{
+class ChatViewController: UIViewController {
     
-    let tableView = UITableView()
-    let messageInputContainer = UIView()
-    let messageTextField = UITextField()
-    let sendButton = UIButton(type: .system)
+    let chatView = ChatView()
     var conversationID: String
     var recipientEmail: String
     var messages: [Message] = []
+
     init(conversationID: String, recipientEmail: String) {
-            self.conversationID = conversationID
-            self.recipientEmail = recipientEmail
-            super.init(nibName: nil, bundle: nil)
-        }
+        self.conversationID = conversationID
+        self.recipientEmail = recipientEmail
+        super.init(nibName: nil, bundle: nil)
+    }
 
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
+
+    override func loadView() {
+        self.view = chatView
+    }
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
         title = recipientEmail
-        setupTableView()
-        setupInputComponents()
         setupKeyboardObservers()
-        tableView.register(MessageCell.self, forCellReuseIdentifier: "MessageCell")
+        chatView.tableView.register(MessageCell.self, forCellReuseIdentifier: "MessageCell")
         observeMessages()
     }
 
-    // MARK: - Table View Setup
-    func setupTableView() {
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        tableView.dataSource = self
-        view.addSubview(tableView)
-        view.addSubview(messageInputContainer)
-        
-        NSLayoutConstraint.activate([
-            tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            tableView.bottomAnchor.constraint(equalTo: messageInputContainer.topAnchor)
-        ])
-    }
-
-    // MARK: - Input View Setup
-    func setupInputComponents() {
-        // Container view
-        messageInputContainer.backgroundColor = UIColor(white: 0.95, alpha: 1)
-        messageInputContainer.translatesAutoresizingMaskIntoConstraints = false
-        
-
-        // Text field
-        messageTextField.placeholder = "Type a message"
-        messageTextField.borderStyle = .roundedRect
-        messageTextField.translatesAutoresizingMaskIntoConstraints = false
-        messageInputContainer.addSubview(messageTextField)
-
-        // Send button
-        sendButton.setTitle("Send", for: .normal)
-        sendButton.setTitleColor(.systemBlue, for: .normal)
-        sendButton.titleLabel?.font = .boldSystemFont(ofSize: 16)
-        sendButton.translatesAutoresizingMaskIntoConstraints = false
-        sendButton.addTarget(self, action: #selector(sendPressed), for: .touchUpInside)
-        messageInputContainer.addSubview(sendButton)
-
-        NSLayoutConstraint.activate([
-            // Container constraints
-            messageInputContainer.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            messageInputContainer.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            messageInputContainer.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
-            messageInputContainer.heightAnchor.constraint(equalToConstant: 60),
-
-            // TextField constraints
-            messageTextField.leadingAnchor.constraint(equalTo: messageInputContainer.leadingAnchor, constant: 16),
-            messageTextField.centerYAnchor.constraint(equalTo: messageInputContainer.centerYAnchor),
-            messageTextField.heightAnchor.constraint(equalToConstant: 40),
-
-            // Send button
-            sendButton.leadingAnchor.constraint(equalTo: messageTextField.trailingAnchor, constant: 8),
-            sendButton.trailingAnchor.constraint(equalTo: messageInputContainer.trailingAnchor, constant: -16),
-            sendButton.centerYAnchor.constraint(equalTo: messageInputContainer.centerYAnchor),
-            sendButton.widthAnchor.constraint(equalToConstant: 60),
-
-            // TextField trailing limited by button
-            messageTextField.trailingAnchor.constraint(equalTo: sendButton.leadingAnchor, constant: -8),
-        ])
-    }
-    @objc func sendPressed() {
-        guard let text = messageTextField.text, !text.isEmpty else { return }
+     @objc func sendPressed() {
+         guard let text = chatView.messageTextField.text, !text.isEmpty else { return }
         guard let currentUserID = Auth.auth().currentUser?.uid else {
             print("❌ No user logged in")
             return
@@ -119,7 +60,7 @@ class ChatViewController: UIViewController{
                     return
                 }
 
-                self.messageTextField.text = ""
+                self.chatView.messageTextField.text = ""
             }
 
         // Update conversation metadata (last message & timestamp)
@@ -131,11 +72,7 @@ class ChatViewController: UIViewController{
             ], merge: true)
     }
     
-    func scrollToBottom() {
-        let indexPath = IndexPath(row: messages.count - 1, section: 0)
-        tableView.scrollToRow(at: indexPath, at: .bottom, animated: true)
-    }
-
+    
     // MARK: - Keyboard Handling
     func setupKeyboardObservers() {
         NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow),
@@ -177,8 +114,8 @@ class ChatViewController: UIViewController{
                 } ?? []
                 
                 DispatchQueue.main.async {
-                    self.tableView.reloadData()
-                    self.scrollToBottom()
+                    self.chatView.tableView.reloadData()
+                   
                 }
             }
     }
