+++ Chatty/Controller/LoginViewController.swift
@@ -6,106 +6,28 @@
 //
 import UIKit
 import FirebaseAuth
-import FirebaseFirestore
 
 class LoginViewController: UIViewController {
     
-    var backgroundView = UIView()
-    var titleLabel = UILabel()
-    var emailTextField = UITextField()
-    var passwordTextField = UITextField()
-    var loginButton = UIButton()
-    var registerButton = UIButton()
-    
+    let loginView = LoginView()
+
+    override func loadView() {
+        self.view = loginView
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        setUpView()
-        applyConstraints()
+        setUpActions()
     }
-    
-    func setUpView() {
-        
-        backgroundView.backgroundColor = UIColor(red: 37/255, green: 211/255, blue: 102/255, alpha: 1)
-        backgroundView.translatesAutoresizingMaskIntoConstraints = false
-        
-        titleLabel.text = "Chatty"
-        titleLabel.font = UIFont.boldSystemFont(ofSize: 34)
-        titleLabel.textColor = .white
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        
-        emailTextField.placeholder = "Email"
-        emailTextField.keyboardType = .emailAddress
-        emailTextField.autocapitalizationType = .none
-        emailTextField.backgroundColor = .white
-        emailTextField.layer.cornerRadius = 8
-        emailTextField.translatesAutoresizingMaskIntoConstraints = false
-        
-        passwordTextField.placeholder = "Password"
-        passwordTextField.isSecureTextEntry = true
-        passwordTextField.backgroundColor = .white
-        passwordTextField.layer.cornerRadius = 8
-        passwordTextField.translatesAutoresizingMaskIntoConstraints = false
-        
-        loginButton.setTitle("Log In", for: .normal)
-        loginButton.backgroundColor = UIColor(named: "#25D366")
-        loginButton.setTitleColor(.white, for: .normal)
-        loginButton.titleLabel?.font = .boldSystemFont(ofSize: 16)
-        loginButton.layer.cornerRadius = 8
-        loginButton.translatesAutoresizingMaskIntoConstraints = false
-        loginButton.addTarget(self, action: #selector(handleLogin), for: .touchUpInside)
-        
-        registerButton.setTitle("Don't have an account? Register", for: .normal)
-        registerButton.setTitleColor(.white, for: .normal)
-        registerButton.titleLabel?.font = .systemFont(ofSize: 14)
-        registerButton.translatesAutoresizingMaskIntoConstraints = false
-        registerButton.addTarget(self, action: #selector(handleRegister), for: .touchUpInside)
-        
-        view.addSubview(backgroundView)
-        view.addSubview(titleLabel)
-        view.addSubview(emailTextField)
-        view.addSubview(passwordTextField)
-        view.addSubview(loginButton)
-        view.addSubview(registerButton)
+
+    private func setUpActions() {
+        loginView.loginButton.addTarget(self, action: #selector(handleLogin), for: .touchUpInside)
+        loginView.registerButton.addTarget(self, action: #selector(handleRegister), for: .touchUpInside)
     }
-    
-    func applyConstraints() {
-        NSLayoutConstraint.activate([
-            
-            backgroundView.topAnchor.constraint(equalTo: view.topAnchor),
-            backgroundView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            backgroundView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            backgroundView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            
-            titleLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 60),
-            titleLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            
-            emailTextField.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 40),
-            emailTextField.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 30),
-            emailTextField.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -30),
-            emailTextField.heightAnchor.constraint(equalToConstant: 44),
-            
-            passwordTextField.topAnchor.constraint(equalTo: emailTextField.bottomAnchor, constant: 20),
-            passwordTextField.leadingAnchor.constraint(equalTo: emailTextField.leadingAnchor),
-            passwordTextField.trailingAnchor.constraint(equalTo: emailTextField.trailingAnchor),
-            passwordTextField.heightAnchor.constraint(equalToConstant: 44),
-            
-            loginButton.topAnchor.constraint(equalTo: passwordTextField.bottomAnchor, constant: 30),
-            loginButton.leadingAnchor.constraint(equalTo: emailTextField.leadingAnchor),
-            loginButton.trailingAnchor.constraint(equalTo: emailTextField.trailingAnchor),
-            loginButton.heightAnchor.constraint(equalToConstant: 50),
-            
-            registerButton.topAnchor.constraint(equalTo: loginButton.bottomAnchor, constant: 20),
-            registerButton.centerXAnchor.constraint(equalTo: view.centerXAnchor)
-        ])
-    }
-    
-    
-    
-    // MARK: - Actions
-    
+
     @objc private func handleLogin() {
-        guard let email = emailTextField.text, !email.isEmpty,
-              let password = passwordTextField.text, !password.isEmpty else {
+        guard let email = loginView.emailTextField.text, !email.isEmpty,
+              let password = loginView.passwordTextField.text, !password.isEmpty else {
             print("❌ Email or password empty")
             return
         }
@@ -120,10 +42,9 @@ class LoginViewController: UIViewController {
             }
         }
     }
+
     @objc private func handleRegister() {
         let registerVC = RegisterViewController()
         navigationController?.pushViewController(registerVC, animated: true)
     }
-    
 }
-
