+++ Chatty/Controller/RegisterViewController.swift
@@ -10,103 +10,36 @@ import FirebaseFirestore
 
 class RegisterViewController: UIViewController {
     
-    var backgroundView = UIView()
-    var titleLabel = UILabel()
-    var emailTextField = UITextField()
-    var passwordTextField = UITextField()
-    var registerButton = UIButton()
+    let registerView = RegisterView()
+    override func loadView() {
+        self.view = registerView
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        setUpView()
-        applyConstaints()
+        setUpActions()
     }
     
-    func setUpView(){
-        backgroundView.backgroundColor = UIColor(red: 37/255, green: 211/255, blue: 102/255, alpha: 1)
-        backgroundView.translatesAutoresizingMaskIntoConstraints = false
-        
-        titleLabel.text = "Register"
-        titleLabel.font = UIFont.boldSystemFont(ofSize: 30)
-        titleLabel.textColor = .white
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        
-        emailTextField.placeholder = "Email"
-        emailTextField.keyboardType = .emailAddress
-        emailTextField.backgroundColor = .white
-        emailTextField.layer.cornerRadius = 8
-        emailTextField.autocapitalizationType = .none
-        emailTextField.translatesAutoresizingMaskIntoConstraints = false
-        
-        passwordTextField.placeholder = "Password"
-        passwordTextField.isSecureTextEntry = true
-        passwordTextField.backgroundColor = .white
-        passwordTextField.layer.cornerRadius = 8
-        passwordTextField.translatesAutoresizingMaskIntoConstraints = false
-        
-        registerButton.setTitle("Create Account", for: .normal)
-        registerButton.backgroundColor = UIColor(named: "#25D366")
-        registerButton.setTitleColor(.white, for: .normal)
-        registerButton.titleLabel?.font = .boldSystemFont(ofSize: 16)
-        registerButton.layer.cornerRadius = 8
-        registerButton.translatesAutoresizingMaskIntoConstraints = false
-        registerButton.addTarget(self, action: #selector(handleRegister), for: .touchUpInside)
-        
-        view.addSubview(backgroundView)
-        view.addSubview(titleLabel)
-        view.addSubview(emailTextField)
-        view.addSubview(passwordTextField)
-        view.addSubview(registerButton)
-        
-    }
-    
-    func applyConstaints(){
-        
-        NSLayoutConstraint.activate([
-            
-            backgroundView.topAnchor.constraint(equalTo: view.topAnchor),
-            backgroundView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            backgroundView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            backgroundView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-        
-            titleLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 60),
-            titleLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            
-            
-            emailTextField.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 20),
-            emailTextField.leadingAnchor.constraint(equalTo: view.leadingAnchor,constant: 30),
-            emailTextField.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -30),
-            emailTextField.heightAnchor.constraint(equalToConstant: 44),
-            
-            passwordTextField.topAnchor.constraint(equalTo: emailTextField.bottomAnchor, constant: 20),
-            passwordTextField.leadingAnchor.constraint(equalTo: emailTextField.leadingAnchor),
-            passwordTextField.trailingAnchor.constraint(equalTo: emailTextField.trailingAnchor),
-            passwordTextField.heightAnchor.constraint(equalToConstant: 44),
-            
-            registerButton.topAnchor.constraint(equalTo: passwordTextField.bottomAnchor, constant: 30),
-            registerButton.leadingAnchor.constraint(equalTo: emailTextField.leadingAnchor),
-            registerButton.trailingAnchor.constraint(equalTo: emailTextField.trailingAnchor),
-            registerButton.heightAnchor.constraint(equalToConstant: 50),
-        
-        ])
+    private func setUpActions() {
+        registerView.registerButton.addTarget(self, action: #selector(handleRegister), for: .touchUpInside)
     }
     
     // MARK: - Actions
     
     
     @objc private func handleRegister() {
-        guard let email = emailTextField.text, !email.isEmpty,
-              let password = passwordTextField.text, !password.isEmpty else {
+        guard let email = registerView.emailTextField.text, !email.isEmpty,
+              let password = registerView.passwordTextField.text, !password.isEmpty else {
             print("❌ Email or password empty")
             return
         }
-
+        
         Auth.auth().createUser(withEmail: email, password: password) { result, error in
             if let error = error {
                 print("❌ Registration failed:", error.localizedDescription)
             } else if let user = result?.user {
                 print("✅ User registered:", user.uid)
-
+                
                 // SAVE USER TO FIRESTORE
                 let db = Firestore.firestore()
                 db.collection("users").document(user.uid).setData([
@@ -116,11 +49,15 @@ class RegisterViewController: UIViewController {
                         print("❌ Failed to save user to Firestore:", error.localizedDescription)
                     } else {
                         print("✅ User saved to Firestore")
-                        // Optional: go to messages screen
+                        
+                        // 🚀 Navigate to Messages screen
+                        DispatchQueue.main.async {
+                            let messagesListVC = MessagesListViewController()
+                            self.navigationController?.pushViewController(messagesListVC, animated: true)
+                        }
                     }
                 }
             }
         }
     }
 }
-
