+++ Chatty/View/LoginView.swift
@@ -1,0 +1,102 @@
+//
+//  LoginView.swift
+//  Chatty
+//
+//  Created by Emmanuel Pena on 7/13/25.
+//
+import UIKit
+
+class LoginView: UIView {
+    
+    // MARK: - UI Elements
+    let backgroundView = UIView()
+    let titleLabel = UILabel()
+    let emailTextField = UITextField()
+    let passwordTextField = UITextField()
+    let loginButton = UIButton()
+    let registerButton = UIButton()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setUpView()
+        applyConstraints()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    // MARK: - View Setup
+    private func setUpView() {
+        backgroundView.backgroundColor = UIColor(red: 37/255, green: 211/255, blue: 102/255, alpha: 1)
+        backgroundView.translatesAutoresizingMaskIntoConstraints = false
+
+        titleLabel.text = "Chatty"
+        titleLabel.font = UIFont.boldSystemFont(ofSize: 34)
+        titleLabel.textColor = .white
+        titleLabel.translatesAutoresizingMaskIntoConstraints = false
+
+        emailTextField.placeholder = "Email"
+        emailTextField.keyboardType = .emailAddress
+        emailTextField.autocapitalizationType = .none
+        emailTextField.backgroundColor = .white
+        emailTextField.layer.cornerRadius = 8
+        emailTextField.translatesAutoresizingMaskIntoConstraints = false
+
+        passwordTextField.placeholder = "Password"
+        passwordTextField.isSecureTextEntry = true
+        passwordTextField.backgroundColor = .white
+        passwordTextField.layer.cornerRadius = 8
+        passwordTextField.translatesAutoresizingMaskIntoConstraints = false
+
+        loginButton.setTitle("Log In", for: .normal)
+        loginButton.backgroundColor = UIColor(red: 37/255, green: 211/255, blue: 102/255, alpha: 1)
+        loginButton.setTitleColor(.white, for: .normal)
+        loginButton.titleLabel?.font = .boldSystemFont(ofSize: 16)
+        loginButton.layer.cornerRadius = 8
+        loginButton.translatesAutoresizingMaskIntoConstraints = false
+
+        registerButton.setTitle("Don't have an account? Register", for: .normal)
+        registerButton.setTitleColor(.white, for: .normal)
+        registerButton.titleLabel?.font = .systemFont(ofSize: 14)
+        registerButton.translatesAutoresizingMaskIntoConstraints = false
+
+        // Add subviews
+        addSubview(backgroundView)
+        addSubview(titleLabel)
+        addSubview(emailTextField)
+        addSubview(passwordTextField)
+        addSubview(loginButton)
+        addSubview(registerButton)
+    }
+
+    private func applyConstraints() {
+        NSLayoutConstraint.activate([
+            backgroundView.topAnchor.constraint(equalTo: topAnchor),
+            backgroundView.bottomAnchor.constraint(equalTo: bottomAnchor),
+            backgroundView.leadingAnchor.constraint(equalTo: leadingAnchor),
+            backgroundView.trailingAnchor.constraint(equalTo: trailingAnchor),
+            
+            titleLabel.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor, constant: 60),
+            titleLabel.centerXAnchor.constraint(equalTo: centerXAnchor),
+            
+            emailTextField.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 40),
+            emailTextField.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 30),
+            emailTextField.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -30),
+            emailTextField.heightAnchor.constraint(equalToConstant: 44),
+            
+            passwordTextField.topAnchor.constraint(equalTo: emailTextField.bottomAnchor, constant: 20),
+            passwordTextField.leadingAnchor.constraint(equalTo: emailTextField.leadingAnchor),
+            passwordTextField.trailingAnchor.constraint(equalTo: emailTextField.trailingAnchor),
+            passwordTextField.heightAnchor.constraint(equalToConstant: 44),
+            
+            loginButton.topAnchor.constraint(equalTo: passwordTextField.bottomAnchor, constant: 30),
+            loginButton.leadingAnchor.constraint(equalTo: emailTextField.leadingAnchor),
+            loginButton.trailingAnchor.constraint(equalTo: emailTextField.trailingAnchor),
+            loginButton.heightAnchor.constraint(equalToConstant: 50),
+            
+            registerButton.topAnchor.constraint(equalTo: loginButton.bottomAnchor, constant: 20),
+            registerButton.centerXAnchor.constraint(equalTo: centerXAnchor)
+        ])
+    }
+}
